rename directory shared/api
